requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.94.1
gunicorn==22.0.0
python-dotenv==1.0.1
pdfplumber==0.11.0
reportlab==4.2.5
sib-api-v3-sdk==7.6.0
# sqlite3 is included in Python's standard library — no install needed
